Fixed a bunch of things, but mainly focused on getting the tilemap to load. There were a lot of silly errors, from accidentally reading in comments lines to getting my rows and columns switched...
Sprite loading/drawing and Ribbon loading/drawing are currently disabled, so I
could focus on getting at least the tile map to draw.

Full screen mode works surprisingly well (tested it once on a whim, was
astonished that I did it right), but I forgot to include the window border in
my drawing calculations. It's been compensated for in the status bar, but I'll
need to factor it in with the rest of the drawing, when I enable it.